android: Support per-app language
Support per-app language settings for Android 13+ users. This is done by
adding the `resources.properties` file with the default locale in
`androidMain/res` and setting `generateLocaleConfig = true`. However
because this is a CMP app we have to add place holder string resources
in `androidMain/res` (e.g `values-de`) so that the correct languages are
suggested in the system UI, even though the actual translations live in
`commonMain/composeResources`. This enables users to set the language
for this specific app in their system/app settings instead of using the
system language (or a in-app language picker (todo)).