android: Improve Barcode Scanner
On android we can limit selectable cameras to improve startup latencey.
Manual testing showed that a 4:3 aspect ratio with higher resolution
than Ml Kit image analysys default (480x..something) but lower than
cameras max resolution yields the best / fastest results. It might be
interesting to see if we can increase the resolution when the user wants
to zoom.

I also fixed some small issues in the `QrCodeAnalyzer` that got in
because of vibe coding, the `processing` atomic boolean is not needed as
`analyze` is called squentially and is blocked until an image is closed.